Added more error checks for leaving and joining a course

+++ application/model.py
@@ -370,12 +370,16 @@ def join_course(username, course_code):
     conn = get_db()
     cur = conn.cursor()
     
-    cur.execute(
-        "INSERT INTO enrollments (username, course_code) "
-        "VALUES (%s, %s) ",
-        (username, course_code)
-    )
-    conn.commit()
+    check = check_course(username, course_code)
+    if(not check["is_joined"]):
+        cur.execute(
+            "INSERT INTO enrollments (username, course_code) "
+            "VALUES (%s, %s) ",
+            (username, course_code)
+        )
+        conn.commit()
+        return {"response" : "success"}
+    return {"response" : "error"}
 
 def get_all_courses_user(username):
     """Get all courses for user that have not been joined yet."""
@@ -398,15 +402,33 @@ def get_all_courses_user(username):
 
     return context
 
+def check_course(username, course_code):
+    conn = get_db()
+    cur = conn.cursor()
+    
+    cur.execute(
+        "SELECT 1 FROM enrollments WHERE username = %s AND course_code = %s",
+        (username, course_code)
+    )
+    
+    enrollment_exists = cur.fetchone()
+    return {"is_joined": enrollment_exists is not None}
+
+
 def drop_course(username, course_code):
     """Allow a user to drop a course."""
     conn = get_db()
     cur = conn.cursor()
-    cur.execute(
-        "DELETE FROM enrollments WHERE username = %s AND course_code = %s",
-        (username, course_code)
-    )
-    conn.commit()
+
+    check = check_course(username, course_code)
+    if(check["is_joined"]):
+        cur.execute(
+            "DELETE FROM enrollments WHERE username = %s AND course_code = %s",
+            (username, course_code)
+        )
+        conn.commit()
+        return {"response" : "success"}
+    return {"response" : "error"}
 
 class InvalidUsage(Exception):
     """Custom exception class for invalid usage of API."""
